requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.3.5
pandas_profiling==3.1.0
Pillow==9.1.0
plotly==5.7.0
seaborn==0.11.2
streamlit==1.8.1
streamlit_pandas_profiling==0.1.3
